refactor(api-headless-cms-utils-os): inline bare functions into DI implementations
Move body.ts logic into BodyBuilderImpl.build(), exec.ts logic into
ExecFilteringImpl.execute(). Move sort.ts, applyFiltering.ts, populated.ts,
values.ts as helpers inside their respective feature folders. Delete old
files: body.ts, sort.ts, exec.ts, path.ts, transformValueForSearch.ts,
plugins/operator.ts, filtering/applyFiltering.ts, filtering/populated.ts,
filtering/values.ts. Keep filtering/index.ts as thin compatibility shim
for ddb-es test imports. Only shared utilities remain in operations/entry/
elasticsearch/ (fields, initialQuery, keyword, fullTextSearch, types, etc).
