+++ packages/api-headless-cms-ddb-es/__tests__/filtering/mocks/plugins.ts
@@ -1,14 +1,13 @@
-import { createOperatorPluginList } from "@webiny/api-headless-cms-utils-os/operations/entry/elasticsearch/plugins/operator";
 import type { OpenSearchQueryBuilderOperators } from "@webiny/api-headless-cms-utils-os/operations/entry/elasticsearch/types";
 import { CmsEntryOpenSearchValueSearchRegistry } from "@webiny/api-headless-cms-utils-os/features/CmsEntryOpenSearchValueSearch";
 import { CmsEntryOpenSearchFilterRegistry } from "@webiny/api-headless-cms-utils-os/features/CmsEntryOpenSearchFilter";
+import { CmsEntryOpenSearchOperatorList } from "@webiny/api-headless-cms-utils-os/features/CmsEntryOpenSearchOperatorList";
 import { createTestContainer } from "~tests/helpers/createTestContainer";
-import { OpenSearchQueryBuilderOperatorRegistry } from "@webiny/api-opensearch/exports/api/opensearch";
 
 export const buildElasticsearchOperatorPlugins = () => {
     const testContainer = createTestContainer();
-    const registry = testContainer.resolve(OpenSearchQueryBuilderOperatorRegistry);
-    return createOperatorPluginList({ registry });
+    const operatorList = testContainer.resolve(CmsEntryOpenSearchOperatorList);
+    return operatorList.getAll();
 };
 
 export interface Plugins {
